some cleanup and renaming
