--- src/login/index.js
@@ -1,9 +1,0 @@
-import LoginComponent from './LoginComponent.vue';
-
-const LoginPlugin = {
-  install(Vue, options) {
-    Vue.component(LoginComponent.name, LoginComponent)
-  }
-};
-
-export default LoginPlugin;
