+++ src/spa/router/index.js
@@ -10,11 +10,6 @@ Vue.use(Router)
 export default new Router({
   routes: [
 
-    // {
-    //   path: '/',
-    //   name: 'Login',
-    //   component: Login
-    // },
     {
       path: '/',
       name: 'Calendar',
